Removed subdomains from legitimate domain requests
This addition is a bandaid fix for an issue where subdomains of legitimate domains will trigger the alert. This fix doesn't address country-code TLDs, but those should not be a problem unless the user inserts them into the legit_domains variable.

This also adds the legitimate domain to the notice in the $sub field.